requirements.txt
python-dotenv==1.0.0
pafy==0.5.5
youtube-dl==2021.12.17
nextcord==2.5.0
wavelinkcord==2.5.1
